tad implementado
